+++ extension/op_registry_loading.py
@@ -214,3 +214,15 @@ def process_registry(context, registry):
         type=component_container,
         override={"LIBRARY_OVERRIDABLE"},
     )
+    bpy.types.Camera.skein_two = bpy.props.CollectionProperty(
+        type=component_container,
+        override={"LIBRARY_OVERRIDABLE"},
+    )
+    bpy.types.Light.skein_two = bpy.props.CollectionProperty(
+        type=component_container,
+        override={"LIBRARY_OVERRIDABLE"},
+    )
+    bpy.types.Collection.skein_two = bpy.props.CollectionProperty(
+        type=component_container,
+        override={"LIBRARY_OVERRIDABLE"},
+    )